Rename PL train entry point
